Adicionado metodos Em aluno dao, e mudança do tipo primitivo int para long

+++ src/main/java/leonardo/barbosa/dao/DaoInterface.java
@@ -10,8 +10,8 @@ public interface DaoInterface {
 
     List<Object> listar();
 
-    Object buscarPorId(int id);
+    Object buscarPorId(Long id);
 
-    boolean deletar(int id);
+    boolean deletar(Long id);
 
 }
